backend/requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
duckdb==0.9.2

# Rate limiting
slowapi==0.1.9

# Environment management
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Optional: Caching
cachetools==5.3.2

# Data processing tools (for scripts in tools/)
pandas==2.1.3
pyarrow==14.0.1
requests==2.31.0
tqdm==4.66.1

# Bluesky collection (optional)
atproto==0.0.45